+++ NEWS.md
@@ -14,6 +14,11 @@
     * `japan_climate` to `climate_jp`
     * `world_climate` to `climate_world`
 
+# clidatajp 0.5.1
+
+* Add graceful fail to match the CRAN policy: 
+    'Packages which use Internet resources should fail gracefully with an informative message if the resource is not available or has changed (and not give a check warning nor error).'
+
 # clidatajp 0.5.0
 
 * First release